Add files via upload
Multiple Classifier Output
===========================
1. hw_1_data_set_check_multiple_topic.py 
Using this we check is there any multiple topic is labeled in dataset or not.
However, we cannot find any multiple labeled data in the dataset.
So we use probability distribution for multiple classifiers.
Since the dataset does not contain multiple labels, the performance evaluation is based on labeled only. 
Accuracy score:  0.829612220917 

2. hw_1_NaiveBayes_Multiple_Output.py
All 10 classifiers for all 10 types of topics.
output_final_1.txt contains the testing output for 10 classifiers.